Push to main for the 1.13.0 version (#74)
* Included support for oscar-cli cluster status (#67)

* Included support for /system/status

oscar-cli cluster status

* status update

* update readme

---------


* Include support for cluster status in the TUI (#68)

* Included support for /system/status

oscar-cli cluster status

* Included support for /system/status in the TUI

* Included support for cluster status in the tui

via oscar-cli interactive

---------


* Refactored TUI & Included Search Functionality (#70)

* Splitted TUI

Split the TUI into focused files so pkg/tui/app.go is now mostly the UI wiring/entrypoint while helpers live alongside it:

+ pkg/tui/state.go holds constants/enums and uiState/bucket object structs; formatters.go and table_helpers.go hold pure helpers.
+ View/controller logic now lives in services_view.go, buckets_view.go, and bucket_objects.go;
+ cross-cutting UI helpers went to dialogs.go, search.go, and auto_refresh.go.
+ app.go cleaned up accordingly, with redundant definitions removed and imports trimmed.

* Improved search functionality

+ Implemented support for search via "/" across Services, Buckets, Clusters and Details panels, including substrings.

* resolve conflict

* resolve conflict

---------


* Improved Log Management in the TUI (#71)

* Splitted TUI

Split the TUI into focused files so pkg/tui/app.go is now mostly the UI wiring/entrypoint while helpers live alongside it:

+ pkg/tui/state.go holds constants/enums and uiState/bucket object structs; formatters.go and table_helpers.go hold pure helpers.
+ View/controller logic now lives in services_view.go, buckets_view.go, and bucket_objects.go;
+ cross-cutting UI helpers went to dialogs.go, search.go, and auto_refresh.go.
+ app.go cleaned up accordingly, with redundant definitions removed and imports trimmed.

* Improved search functionality

+ Implemented support for search via "/" across Services, Buckets, Clusters and Details panels, including substrings.

* Improved log management in the TUI

---------


* fix invocation (#73)

* Allow to override bucket name to match service name (#72)

This is applied both for oscar-cli hub and oscar-cli apply

---------
